[Feat]#2 - Add UIbutton, UIsearchBar - add extension click map
